src/functions/airtable_sync/requirements.txt
functions-framework>=3.5.0
google-cloud-secret-manager>=2.18.1
requests>=2.31.0